% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSTP.R
\name{CDSTP}
\alias{CDSTP}
\alias{dCDSTP}
\alias{pCDSTP}
\alias{rCDSTP}
\title{Continuous Dual-Stage Two-Phase Model of Selective Attention}
\usage{
dCDSTP(
  rt,
  resp,
  phi = c(0.3, 0.5, 0.5, -0.5, -1, -0.5, 8, 4, 1, 2, 1.3, 1.3, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

pCDSTP(
  rt,
  resp,
  phi = c(0.3, 0.5, 0.5, -0.5, -1, -0.5, 8, 4, 1, 2, 1.3, 1.3, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

rCDSTP(
  n,
  phi = c(0.3, 0.5, 0.5, -0.5, -1, -0.5, 8, 4, 1, 2, 1.3, 1.3, 0, 0, 1),
  dt = 1e-05
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time of process 1
\item relative starting point of process 1
\item relative starting point for process 2
\item ?
\item ?
\item ?
\item ?
\item ?
\item ?
\item ?
\item upper threshold of process 1 (equals negative lower threshold)
\item upper threshold of process 2 (equals negative lower threshold)
\item contamination strength of process 1
\item contamination probability for the lower response of process 1
\item contamination probability for the upper response of process 1
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for the continuous
dual-stage two-phase (CDSTP) model of selective attention by Hübner, Steinhauser,
and Lehle (2010).
}
\examples{
# Probability density function
dCDSTP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
       phi = c(0.3, 0.5, 0.5, -0.5, -1.0, -0.5, 8.0, 4.0, 1.0, 2.0, 1.3, 1.3, 0.0, 0.0, 1.0))

# Cumulative distribution function
pCDSTP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
       phi = c(0.3, 0.5, 0.5, -0.5, -1.0, -0.5, 8.0, 4.0, 1.0, 2.0, 1.3, 1.3, 0.0, 0.0, 1.0))

# Random sampling
rCDSTP(n = 100, phi = c(0.3, 0.5, 0.5, -0.5, -1.0, -0.5, 8.0, 4.0, 1.0, 2.0, 1.3, 1.3,
                        0.0, 0.0, 1.0))
}
\references{
Hübner, R., Steinhauser, M., & Lehle, C. (2010). A dual-stage two-phase model of
selective attention. \emph{Psychological review, 117}(3), 759.

Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
