% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSTM_T.R
\name{CSTM_T}
\alias{CSTM_T}
\alias{dCSTM_T}
\alias{pCSTM_T}
\alias{rCSTM_T}
\title{Custom Time-Dependent Drift Diffusion Model}
\usage{
dCSTM_T(rt, resp, phi, x_res = "default", t_res = "default")

pCSTM_T(rt, resp, phi, x_res = "default", t_res = "default")

rCSTM_T(n, phi, dt = 1e-05)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in your specified order}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for a custom
time-dependent (CSTM_T) drift diffusion model. Run
\code{vignette("guideline", package = "ream")} to see how custom models work.
}
\examples{
# Probability density function
dCSTM_T(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
      phi = c(0.3, 0.5, 1.0, 1.0, 0.75, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
