% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSTM_TX.R
\name{dCSTM_TX_grid}
\alias{dCSTM_TX_grid}
\title{Generate Grid for PDF of Custom Time- and Evidence-Dependent Drift Diffusion Model}
\usage{
dCSTM_TX_grid(rt_max = 10, phi, x_res = "default", t_res = "default")
}
\arguments{
\item{rt_max}{maximal response time <- max(rt)}

\item{phi}{parameter vector in your order}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
such and such
}
\description{
Beschreibung.
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter inference for a wide class of binary evidence accumulation models.
Behavior Research Methods.
}
\author{
Raphael Hartmann & Matthew Murrow
}
