% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMC.R
\name{dDMC}
\alias{dDMC}
\title{PDF of the Diffusion Model for Conflict Tasks}
\usage{
dDMC(
  rt,
  resp,
  phi = c(0.1, 0.5, 0.2, 0.2, 0.05, 2.5, 3, 1, 0.5, 0, 0, 1),
  x_res = "A",
  t_res = "A"
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point (spBias)
\item congruency (1.0: congruent; -1.0: incongruent)
\item peak amplitude of automatic activation (automatic drift rate; amp)
\item time to peak automatic activation (automatic drift rate; tau)
\item shape of automatic process drift rate (automatic drift rate; aaShape)
\item drift rate for controlled process (drc)
\item diffusion rate (typically 1.0; sigm)
\item upper threshold (equals negative lower threshold; bnds)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{rt_max}{maximal response time <- max(rt)}
}
\value{
a list of PDF values, log-PDF values, and the sum of the log-PDFs
}
\description{
Calculation the (Log-)PDF of the diffusion model for conflict tasks (DMC) by Ulrich,
Schröter, Leuthold, and Birngruber (2015).
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

Ulrich, R., Schröter, H., Leuthold, H., & Birngruber, T. (2015). Automatic
and controlled stimulus processing in conflict tasks: Superimposed diffusion
processes and delta functions. \emph{Cognitive psychology, 78}, 148-174.
}
\author{
Raphael Hartmann & Matthew Murrow
}
