% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMC.R
\name{dDMC_grid}
\alias{dDMC_grid}
\title{Generate Grid for PDF of Diffusion Model for Conflict Tasks}
\usage{
dDMC_grid(
  rt_max = 10,
  phi = c(0.3, 0.5, -1, 0.2, 0.05, 2.5, 3, 1, 0.5, 0, 0, 1),
  x_res = "A",
  t_res = "A"
)
}
\arguments{
\item{rt_max}{maximal response time <- max(rt)}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point (spBias)
\item congruency (1.0: congruent; -1.0: incongruent)
\item peak amplitude of automatic activation (automatic drift rate; amp)
\item time to peak automatic activation (automatic drift rate; tau)
\item shape of automatic process drift rate (automatic drift rate; aaShape)
\item drift rate for controlled process (drc)
\item diffusion rate (typically 1.0; sigm)
\item upper threshold (equals negative lower threshold; bnds)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
such and such
}
\description{
Beschreibung.
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter inference for a wide class of binary evidence accumulation models.
Behavior Research Methods.
}
\author{
Raphael Hartmann & Matthew Murrow
}
