% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM.R
\name{dLTM_grid}
\alias{dLTM_grid}
\title{Generate Grid for PDF of the Linear Threshold Model}
\usage{
dLTM_grid(
  rt_max = 10,
  phi = c(0.3, 0.5, 1, 1, 1.5, 1, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)
}
\arguments{
\item{rt_max}{maximal response time <- max(rt)}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item drift rate
\item diffusion rate
\item ?
\item ?
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
such and such
}
\description{
Beschreibung.
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter inference for a wide class of binary evidence accumulation models.
Behavior Research Methods.

Kempe, D., Kleinberg, J., & Tardos, É. (2003, August). Maximizing the spread
of influence through a social network. In Proceedings of the ninth ACM SIGKDD
international conference on Knowledge discovery and data mining (pp. 137-146).
}
\author{
Raphael Hartmann & Matthew Murrow
}
