% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPM.R
\name{DPM}
\alias{DPM}
\alias{dDPM}
\alias{pDPM}
\alias{rDPM}
\title{Dual Process Model}
\usage{
dDPM(
  rt,
  resp,
  phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1, 0.5, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

pDPM(
  rt,
  resp,
  phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1, 0.5, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

rDPM(n, phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1, 0.5, 0, 0, 1), dt = 1e-05)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point (spBias)
\item p_outer
\item p_inner
\item p_target
\item t_s
\item drift rate for controlled process (drc)
\item diffusion rate (typically 1.0; sigm)
\item upper threshold (equals negative lower threshold; bnds)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for the dual process model.
}
\examples{
# Probability density function
dDPM(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1.0, 0.5, 0.0, 0.0, 1.0))

# Cumulative distribution function
pDPM(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1.0, 0.5, 0.0, 0.0, 1.0))

# Random sampling
rDPM(n = 100, phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1.0, 0.5, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
