% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDDM.R
\name{dSDDM_grid}
\alias{dSDDM_grid}
\title{Generate Grid for PDF of the Simple Drift Diffusion Model}
\usage{
dSDDM_grid(
  rt_max = 10,
  phi = c(0.3, 0.5, 1, 1, 0.75, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)
}
\arguments{
\item{rt_max}{maximal response time <- max(rt)}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item drift rate
\item diffusion rate
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
such and such
}
\description{
Beschreibung.
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter inference for a wide class of binary evidence accumulation models.
Behavior Research Methods.

????
}
\author{
Raphael Hartmann & Matthew Murrow
}
