% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGM.R
\name{dUGM}
\alias{dUGM}
\title{PDF of the Urgency Gating Model}
\usage{
dUGM(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 0.5, 0.5, 1, 1.5, 0, 0, 1),
  x_res = "A",
  t_res = "A"
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item ?
\item ?
\item ?
\item ?
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{rt_max}{maximal response time <- max(rt)}
}
\value{
a list of PDF values, log-PDF values, and the sum of the log-PDFs
}
\description{
Calculation the (Log-)PDF of the urgency gating model (UGM) by Cisek, Puskas, and
El-Murr (2009)
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

Cisek, P., Puskas, G. A., & El-Murr, S. (2009). Decisions in changing conditions:
the urgency-gating model. \emph{Journal of Neuroscience, 29}(37), 11560-11571.
}
\author{
Raphael Hartmann & Matthew Murrow
}
