% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTM.R
\name{dWTM}
\alias{dWTM}
\title{PDF of the Weibull Threshold Model}
\usage{
dWTM(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 1, 1.5, 0.2, 0.5, -1, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item drift rate
\item diffusion rate
\item ?
\item ?
\item ?
\item ?
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
a list of PDF values, log-PDF values, and the sum of the log-PDFs
}
\description{
Calculation the (Log-)PDF of the Weibull threshold model (WTM)
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
