% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPM.R
\name{pDPM}
\alias{pDPM}
\title{CDF of the Dual Process Model}
\usage{
pDPM(
  rt,
  resp,
  phi = c(0.25, 0.5, -0.3, -0.3, 0.3, 0.25, 1, 0.5, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point (spBias)
\item p_outer
\item p_inner
\item p_target
\item t_s
\item drift rate for controlled process (drc)
\item diffusion rate (typically 1.0; sigm)
\item upper threshold (equals negative lower threshold; bnds)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
a list of PDF values, log-PDF values, and the sum of the log-PDFs
}
\description{
Calculation the (Log-)CDF of the dual process model.
}
\examples{
# here come some examples
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
