% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{rLM}
\alias{rLM}
\title{Sampling From the Leakage Model}
\usage{
rLM(n, phi = c(0.3, 0.5, 1, 0.5, 1, 0.5, 0, 0, 1), dt = 1e-05)
}
\arguments{
\item{n}{number of samples}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item ?
\item ?
\item diffusion rate
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{dt}{step size of time. We recommend 0.00001 for the WTM}
}
\value{
list of response times (rt) and response threholds (resp)
}
\description{
Sampling from the leakage model (WTM)
}
\examples{
rLM(n = 100, phi = c(0.3, 0.5, 1.0, 0.5, 1.0, 0.5, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
