% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM.R
\name{rLTM}
\alias{rLTM}
\title{Sampling From the Linear Threshold Model}
\usage{
rLTM(n, phi = c(0.3, 0.5, 1, 1, 1.5, 1, 0, 0, 1), dt = 1e-05)
}
\arguments{
\item{n}{number of samples}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item drift rate
\item diffusion rate
\item ?
\item ?
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{dt}{step size of time. We recommend 0.00001 for the LTM}
}
\value{
list of response times (rt) and response threholds (resp)
}
\description{
Sampling from the linear threshold model (LTM)
}
\examples{
rLTM(n = 100, phi = c(0.3, 0.5, 1.0, 1.0, 1.5, 1.0, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

Kempe, D., Kleinberg, J., & Tardos, É. (2003, August). Maximizing the spread
of influence through a social network. In Proceedings of the ninth ACM SIGKDD
international conference on Knowledge discovery and data mining (pp. 137-146).
}
\author{
Raphael Hartmann & Matthew Murrow
}
