% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDMC.R
\name{rRDMC}
\alias{rRDMC}
\title{Sampling From the Revised Diffusion Model for Conflict Tasks}
\usage{
rRDMC(n, phi = c(0.35, 0.5, 7.5, 40, 5, 5, 1, 0.5, 0, 0, 1), dt = 1e-05)
}
\arguments{
\item{n}{number of samples}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point (spBias)
\item ? peak amplitude of automatic activation (automatic drift rate; amp)
\item ? time to peak automatic activation (automatic drift rate; tau)
\item ? shape of automatic process drift rate (automatic drift rate; aaShape)
\item ? drift rate for controlled process (drc)
\item diffusion rate (typically 1.0; sigm)
\item upper threshold (equals negative lower threshold; bnds)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{dt}{step size of time. We recommend 0.00001 for the RDMC}
}
\value{
list of response times (rt) and response threholds (resp)
}
\description{
Sampling from the revised diffusion model for conflict tasks (RDMC) by Lee
and Sewell (2023)
}
\examples{
rRDMC(n = 100, phi = c(0.35, 0.5, 7.5, 40.0, 5.0, 5.0, 1.0, 0.5, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

Lee, P. S., & Sewell, D. K. (2023). A revised diffusion model for conflict tasks.
\emph{Psychonomic Bulletin & Review}, 1-31.
}
\author{
Raphael Hartmann & Matthew Murrow
}
