% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSP.R
\name{rSSP}
\alias{rSSP}
\title{Sampling From the Shrinking Spotlight Model}
\usage{
rSSP(
  n,
  phi = c(0.25, 0.5, 1.5, 0.02, -0.3, -0.3, 0.3, 1.5, 0.5, 1.5, -0.5, 0.5, 0.1, 0.04, 0,
    0, 1),
  dt = 1e-05
)
}
\arguments{
\item{n}{number of samples}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item width of the spotlight at t = 0 (sd_a(0))
\item rate with which the width decreases (r_d)
\item perceptual input strength of outer flanker
\item perceptual input strength of inner flanker
\item perceptual input strength of target
\item lower bound of integral for a_outer (typically fixed to 1.5)
\item lower bound of integral for a_inner (typically fixed to 0.5)
\item upper bound of integral for a_inner (typically fixed to 1.5)
\item lower bound of integral for a_target (typically fixed to -0.5)
\item upper bound of integral for a_target (typically fixed to 0.5)
\item diffusion rate
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{dt}{step size of time. We recommend 0.00001 for the SSP}
}
\value{
list of response times (rt) and response threholds (resp)
}
\description{
Sampling from the shrinking spotlight model (SSP)
}
\examples{
rSSP(n = 100, phi = c(0.25, 0.5, 1.5, 0.02, -0.3, -0.3, 0.3, 1.5, 0.5, 1.5, -0.5, 0.5, 0.1, 0.04, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

White, C. N., Ratcliff, R., & Starns, J. J. (2011). Diffusion models of the flanker
task: Discrete versus gradual attentional selection. \emph{Cognitive psychology,
63}(4), 210-238.
}
\author{
Raphael Hartmann & Matthew Murrow
}
