% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSP.R
\name{SSP}
\alias{SSP}
\alias{dSSP}
\alias{pSSP}
\alias{rSSP}
\title{Shrinking Spotlight Model}
\usage{
dSSP(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 7.5, -1, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75, 1, 0.75, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

pSSP(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 7.5, -1, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75, 1, 0.75, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

rSSP(
  n,
  phi = c(0.3, 0.5, 1, 7.5, -1, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75, 1, 0.75, 0, 0, 1),
  dt = 1e-05
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item width of the spotlight at t = 0 (sd_a(0))
\item rate with which the width decreases (r_d)
\item (c)
\item lower bound of integral for a_target (typically fixed to -0.5)
\item upper bound of integral for a_target (typically fixed to 0.5)
\item upper bound of integral for a_inner (typically fixed to 1.5)
\item perceptual input strength of target
\item perceptual input strength of inner flanker divided by c
\item perceptual input strength of outer flanker divided by c
\item diffusion rate
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for the shrinking
spotlight model (SSP).
}
\examples{
# Probability density function
dSSP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.3, 0.5, 1.0, 7.5, -1.0, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75, 1.0,
             0.75, 0.0, 0.0, 1.0))

# Cumulative distribution function
pSSP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.3, 0.5, 1.0, 7.5, -1.0, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75, 1.0,
             0.75, 0.0, 0.0, 1.0))

# Random sampling
rSSP(n = 100, phi = c(0.3, 0.5, 1.0, 7.5, -1.0, -0.5, 0.5, 1.5, 3.75, 3.75, 3.75,
                      1.0, 0.75, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
