% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGM.R
\name{UGM}
\alias{UGM}
\alias{dUGM}
\alias{pUGM}
\alias{rUGM}
\title{Urgency Gating Model}
\usage{
dUGM(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 0.5, 0.5, 1, 1.5, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

pUGM(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 0.5, 0.5, 1, 1.5, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

rUGM(n, phi = c(0.3, 0.5, 1, 0.5, 0.5, 1, 1.5, 0, 0, 1), dt = 1e-05)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time
\item relative starting point
\item ?
\item ?
\item ?
\item diffusion rate
\item upper threshold (equals negative lower threshold)
\item contamination strength
\item contamination probability for the lower response
\item contamination probability for the upper response
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for the urgency gating
model (UGM) by Cisek, Puskas, and El-Murr (2009).
}
\examples{
# Probability density function
dUGM(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.3, 0.5, 1.0, 0.5, 0.5, 1.0, 1.5, 0.0, 0.0, 1.0))

# Cumulative distribution function
pUGM(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
     phi = c(0.3, 0.5, 1.0, 0.5, 0.5, 1.0, 1.5, 0.0, 0.0, 1.0))

# Random sampling
rUGM(n = 100, phi = c(0.3, 0.5, 1.0, 0.5, 0.5, 1.0, 1.5, 0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.

Cisek, P., Puskas, G. A., & El-Murr, S. (2009). Decisions in changing conditions:
the urgency-gating model. \emph{Journal of Neuroscience, 29}(37), 11560-11571.
}
\author{
Raphael Hartmann & Matthew Murrow
}
