% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDSTP.R
\name{WDSTP}
\alias{WDSTP}
\alias{dWDSTP}
\alias{pWDSTP}
\alias{rWDSTP}
\title{W???? Dual-Stage Two-Phase Model of Selective Attention}
\usage{
dWDSTP(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 1, -1, -0.5, 8, 1, 1, 1, 2, 1.3, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

pWDSTP(
  rt,
  resp,
  phi = c(0.3, 0.5, 1, 1, -1, -0.5, 8, 1, 1, 1, 2, 1.3, 0, 0, 1),
  x_res = "default",
  t_res = "default"
)

rWDSTP(
  n,
  phi = c(0.3, 0.5, 1, 1, -1, -0.5, 8, 1, 1, 1, 2, 1.3, 0, 0, 1),
  dt = 1e-05
)
}
\arguments{
\item{rt}{vector of response times}

\item{resp}{vector of responses ("upper" and "lower")}

\item{phi}{parameter vector in the following order:
\itemize{
\item non-decision time of process 1
\item relative starting point of process 1
\item relative starting point for process 2
\item ?
\item ?
\item ?
\item ?
\item ?
\item ?
\item ?
\item ?
\item upper threshold of process (equals negative lower threshold)
\item contamination strength of process
\item contamination probability for the lower response of process
\item contamination probability for the upper response of process
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}

\item{n}{number of samples}

\item{dt}{step size of time. We recommend 0.00001 (1e-5)}
}
\value{
For the density a list of PDF values, log-PDF values, and the sum of the
log-PDFs, for the distribution function a list of of CDF values, log-CDF values,
and the sum of the log-CDFs, and for the random sampler a list of response
times (rt) and response thresholds (resp).
}
\description{
Density (PDF), distribution function (CDF), and random sampler for the w??????
dual-stage two-phase (WDSTP) model of selective attention by Hübner, Steinhauser,
and Lehle (2010).
}
\examples{
# Probability density function
dWDSTP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
       phi = c(0.3, 0.5, 1.0, 1.0, -1.0, -0.5, 8.0, 1.0, 1.0, 1.0, 2.0, 1.3, 0.0, 0.0, 1.0))

# Cumulative distribution function
pWDSTP(rt = c(1.2, 0.6, 0.4), resp = c("upper", "lower", "lower"),
       phi = c(0.3, 0.5, 1.0, 1.0, -1.0, -0.5, 8.0, 1.0, 1.0, 1.0, 2.0, 1.3, 0.0, 0.0, 1.0))

# Random sampling
rWDSTP(n = 100, phi = c(0.3, 0.5, 1.0, 1.0, -1.0, -0.5, 8.0, 1.0, 1.0, 1.0, 2.0, 1.3,
                        0.0, 0.0, 1.0))
}
\references{
Murrow, M., & Holmes, W. R. (2023). PyBEAM: A Bayesian approach to parameter
inference for a wide class of binary evidence accumulation models.
\emph{Behavior Research Methods}, 1-21.
}
\author{
Raphael Hartmann & Matthew Murrow
}
