% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM.R
\name{dLIM_grid}
\alias{dLIM_grid}
\title{Generate Grid for PDF of the Leaky Integration Model}
\usage{
dLIM_grid(rt_max = 10, phi, x_res = "default", t_res = "default")
}
\arguments{
\item{rt_max}{maximal response time <- max(rt)}

\item{phi}{parameter vector in the following order:
\enumerate{
\item Non-decision time (\eqn{t_{nd}}). Time for non-decision processes such as stimulus
encoding and response execution. Total decision time t is the sum of the decision
and non-decision times.
\item Relative start (\eqn{w}). Sets the start point of accumulation as a ratio of
the two decision thresholds. Related to the absolute start z point via equation
\eqn{z = b_l + w*(b_u - b_l)}.
\item Stimulus strength (\eqn{\mu}). Strength of the stimulus.
\item Log10-leakage (\eqn{log_{10}(L)}). Rate of leaky integration.
\item Noise scale (\eqn{\sigma}). Model scaling parameter.
\item Decision thresholds (\eqn{b}). Sets the location of each decision threshold. The
upper threshold \eqn{b_u} is above 0 and the lower threshold \eqn{b_l} is below 0 such that
\eqn{b_u = -b_l = b}. The threshold separation \eqn{a = 2b}.
\item Contamination (\eqn{g}). Sets the strength of the contamination process. Contamination
process is a uniform distribution \eqn{f_c(t)} where \eqn{f_c(t) = 1/(g_u-g_l)}
if \eqn{g_l <= t <= g_u} and \eqn{f_c(t) = 0} if \eqn{t < g_l} or \eqn{t > g_u}. It is
combined with PDF \eqn{f_i(t)} to give the final combined distribution
\eqn{f_{i,c}(t) = g*f_c(t) + (1-g)*f_i(t)}, which is then output by the program.
If \eqn{g = 0}, it just outputs \eqn{f_i(t)}.
\item Lower bound of contamination distribution (\eqn{g_l}). See parameter \eqn{g}.
\item Upper bound of contamination distribution (\eqn{g_u}). See parameter \eqn{g}.
}}

\item{x_res}{spatial/evidence resolution}

\item{t_res}{time resolution}
}
\value{
list of RTs and corresponding defective PDFs at lower and upper threshold
}
\description{
Generate a grid of response-time values and the corresponding PDF values.
For more details on the model see, for example, \code{\link{dLIM}}.
}
\references{
Busemeyer, J. R., & Townsend, J. T. (1993). Decision field theory: A dynamic-cognitive
approach to decision making in an uncertain environment. \emph{Psychological Review, 100}(3), 432-459.

Usher, M., & McClelland, J. L. (2001). The time course of perceptual choice: The leaky,
competing accumulator model. \emph{Psychological Review, 108}(3), 550-592.

Wang, J.-S., & Donkin, C. (2024). The neural implausibility of the diffusion decision
model doesn’t matter for cognitive psychometrics, but the Ornstein-Uhlenbeck model
is better. \emph{Psychonomic Bulletin & Review}.

Wong, K.-F., & Wang, X.-J. (2006). A Recurrent Network Mechanism of Time Integration
in Perceptual Decisions. \emph{The Journal of Neuroscience, 26}(4), 1314-1328.
}
\author{
Raphael Hartmann & Matthew Murrow
}
